adapted to new prototype for *_write function

+++ c/src/lib/libbsp/m68k/gen68340/include/m340uart.h
@@ -81,7 +81,7 @@ Find_Right_m340_UART_Config(float ChannelA_ReceiverBaudRate, float ChannelA_Tran
 extern rtems_isr InterruptHandler (rtems_vector_number v);
 
 extern int dbugRead (int minor);
-extern int dbugWrite (int minor, const char *buf, int len);
+extern ssize_t dbugWrite (int minor, const char *buf, size_t len);
 
 extern float m340_Baud_Rates_Table[16][2];
