+++ c/src/lib/libbsp/m68k/mcf5206elite/console/console.c
@@ -63,8 +63,8 @@ console_poll_read(int minor)
  * RETURNS:
  *     result code
  */
-static int
-console_interrupt_write(int minor, const char *buf, int len)
+static ssize_t
+console_interrupt_write(int minor, const char *buf, size_t len)
 {
     return mcfuart_interrupt_write(&uart[minor], buf, len);
 }
@@ -80,8 +80,8 @@ console_interrupt_write(int minor, const char *buf, int len)
  * RETURNS:
  *     result code
  */
-static int
-console_poll_write(int minor, const char *buf, int len)
+static ssize_t
+console_poll_write(int minor, const char *buf, size_t len)
 {
     return mcfuart_poll_write(&uart[minor], buf, len);
 }
